requirements.txt
streamlit
numpy
opencv-python-headless
scikit-learn
scikit-image
joblib
Pillow